Added an installer for people running Safari 4
The installer should work for both Leopard and Tiger OSes.  However, I only have access to
a maching running Leopard.  There is nothing to determine whether the user is using Safari 4
in the installer.  Therefore there will be no error if a user running Safari 3 attempts
to run the installer.